code for vggt

+++ vggt/dependency/track_modules/base_track_predictor.py
@@ -136,12 +136,12 @@ class BaseTrackerPredictor(nn.Module):
             if transformer_input.shape[2] < self.transformer_dim:
                 # pad the features to match the dimension
                 pad_dim = self.transformer_dim - transformer_input.shape[2]
-                pad = torch.zeros_like(flows_emb[..., 0:pad_dim])
+                pad = torch.zeros_like(flows_emb[..., 0:pad_dim], dtype=flows_emb.dtype)
                 transformer_input = torch.cat([transformer_input, pad], dim=2)
 
             # 2D positional embed
             # TODO: this can be much simplified
-            pos_embed = get_2d_sincos_pos_embed(self.transformer_dim, grid_size=(HH, WW)).to(query_points.device)
+            pos_embed = get_2d_sincos_pos_embed(self.transformer_dim, grid_size=(HH, WW), dtype=transformer_input.dtype).to(query_points.device)
             sampled_pos_emb = sample_features4d(pos_embed.expand(B, -1, -1, -1), coords[:, 0])
             sampled_pos_emb = rearrange(sampled_pos_emb, "b n c -> (b n) c").unsqueeze(1)
