+++ vggt/utils/geometry.py
@@ -24,11 +24,11 @@ def unproject_depth_map_to_point_map(
         np.ndarray: Batch of 3D world coordinates of shape (S, H, W, 3)
     """
     if isinstance(depth_map, torch.Tensor):
-        depth_map = depth_map.cpu().numpy()
+        depth_map = depth_map.cpu().numpy().astype(float)
     if isinstance(extrinsics_cam, torch.Tensor):
-        extrinsics_cam = extrinsics_cam.cpu().numpy()
+        extrinsics_cam = extrinsics_cam.cpu().numpy().astype(float)
     if isinstance(intrinsics_cam, torch.Tensor):
-        intrinsics_cam = intrinsics_cam.cpu().numpy()
+        intrinsics_cam = intrinsics_cam.cpu().numpy().astype(float)
 
     world_points_list = []
     for frame_idx in range(depth_map.shape[0]):
